requirements.txt
altgraph==0.17
aniso8601==8.0.0
app==0.0.1
app-config==1.0.1
asgiref==3.2.3
boto==2.49.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
codegen==1.0
colorama==0.4.3
config==0.5.0.post0
cycler==0.10.0
dj-database-url==0.5.0
django>=3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==19.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.2.1
numpy==1.18.4
oauthlib==3.1.0
Paginator==0.5.1
paho-mqtt==1.5.0
pandas==1.0.3
pathlib==1.0.1
pefile==2019.4.18
Pillow>=7.1.0
plotly==4.8.1
psycopg2==2.8.5
PyInstaller==3.6
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.13.0
python-nvd3==0.15.0
python-slugify==4.0.0
python-socketio==4.6.0
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.2
qrcode==6.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
Rx==3.1.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
view==0.1
views==0.3
vincent==0.4.4
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
